+++ src/IniFileException.php
@@ -46,6 +46,14 @@ class IniFileException extends \Exception
      */
     const ERR_FILE_NOT_READABLE = 6;
     /**
+     * @var Integer Exception code; The INI file does not exist
+     */
+    const ERR_FILE_NOT_EXIST = 7;
+    /**
+     * @var Integer Exception code; The object is in read only mode and a write operation was attempted
+     */
+    const ERR_READ_ONLY_MODE = 8;
+    /**
      * @var Integer Exception code; No exception code was provided in the throw statement, or no pre-defined codes match
      *   the scenario
      */